requirements.txt
fastapi
cryptography
pynacl
pycryptodome
uvicorn